+++ Aeronautikal/src/com/tikal/aeronautikal/controller/CategoriaController.java
@@ -17,6 +17,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
 import com.tikal.aeronautikal.dao.CategoriaDao;
+import com.tikal.aeronautikal.dao.PerfilDAO;
+import com.tikal.aeronautikal.dao.SessionDao;
+import com.tikal.aeronautikal.dao.UsuarioDao;
 import com.tikal.aeronautikal.entity.Categoria;
 import com.tikal.aeronautikal.service.CategoriaService;
 
@@ -34,6 +37,18 @@ public class CategoriaController {
 	 @Autowired
 	 @Qualifier("categoriaDao")
 	 CategoriaDao categoriaDao;
+	 
+	 @Autowired
+	 @Qualifier("sessionDao")
+	 SessionDao sessionDao;
+	 
+	@Autowired
+	@Qualifier ("usuarioDao")
+	UsuarioDao usuarioDao;
+
+		
+	@Autowired
+	PerfilDAO perfilDAO; 
 	 
 
 	 @RequestMapping(value={"/prueba"},method = RequestMethod.GET)
@@ -66,24 +81,28 @@ public class CategoriaController {
 	 /////////////////////////////////////////////////////********************************************************
 
 	 
-	 @RequestMapping(value = {"/add"}, method = RequestMethod.POST, produces = "application/json", consumes = "application/json") 
-	   public void addCategoria(HttpServletResponse response, HttpServletRequest request, @RequestBody String json) throws IOException{
-	    	  System.out.println("si entra al add Categoria por POST"+json);
-	        try {
-	        	AsignadorDeCharset.asignar(request, response);
-	        	 System.out.println("request......."+request);
-	        	 System.out.println("request......."+response);
-	        	 Categoria c =(Categoria) JsonConvertidor.fromJson(json,Categoria.class);
-	        	// System.out.println("el nuevo objeto: "+orden );
-	        	//pegar el valor de empresa, aeronave y contacato
-	        	//cmp.setD_pendientes(50);//aqui va funcion para calcular cuantas piezas pendientes hay de cada componente
-	        	//orden.setFolio(1111);
-	        	categoriaDao.save(c);	            
-	        } catch (RuntimeException ignored) {
-	        	ignored.printStackTrace();
-	            // getUniqueEntity should throw exception
-	        }
-	       
+	 @RequestMapping(value = {"/add/{userName}"}, method = RequestMethod.POST, produces = "application/json", consumes = "application/json") 
+	   public void addCategoria(HttpServletResponse response, HttpServletRequest request, 
+			   @RequestBody String json,  @PathVariable String userName) throws IOException{
+	       System.out.println("si entra al add Categoria por POST"+json);
+	       if(SesionController.verificarPermiso2(request, usuarioDao, perfilDAO, 24, sessionDao,userName)){  
+		        try {
+		        	AsignadorDeCharset.asignar(request, response);
+		        	 System.out.println("request......."+request);
+		        	 System.out.println("request......."+response);
+		        	 Categoria c =(Categoria) JsonConvertidor.fromJson(json,Categoria.class);
+		        	// System.out.println("el nuevo objeto: "+orden );
+		        	//pegar el valor de empresa, aeronave y contacato
+		        	//cmp.setD_pendientes(50);//aqui va funcion para calcular cuantas piezas pendientes hay de cada componente
+		        	//orden.setFolio(1111);
+		        	categoriaDao.save(c);	            
+		        } catch (RuntimeException ignored) {
+		        	ignored.printStackTrace();
+		            // getUniqueEntity should throw exception
+		        }
+	       }else{
+				response.sendError(403);
+			}
 	    }
 	 
 	   /////////////////////////////////////////////////////////////////////////////////////////**********************
@@ -101,26 +120,33 @@ public class CategoriaController {
 	     
 	   
 	   
-	   @RequestMapping(value = {"/delete/{id}" }, method = RequestMethod.POST)
-	   public void deleteCategoria(HttpServletResponse response, HttpServletRequest request, @PathVariable Long id)
-			   throws IOException {
+	   @RequestMapping(value = {"/delete/{id}/{userName}" }, method = RequestMethod.POST)
+	   public void deleteCategoria(HttpServletResponse response, HttpServletRequest request, 
+			   @PathVariable Long id, @PathVariable String userName)   throws IOException {
 		   System.out.println("si esta en delete"+id);
-		   categoriaDao.delete(categoriaDao.consult(id));
-		   System.out.println("Categoria eliminado....");
-		   response.getWriter().println("ok");
-		   
+		   if(SesionController.verificarPermiso2(request, usuarioDao, perfilDAO, 26, sessionDao,userName)){  
+			   categoriaDao.delete(categoriaDao.consult(id));
+			   System.out.println("Categoria eliminado....");
+			   response.getWriter().println("ok");
+		   }else{
+				response.sendError(403);
+			}
 	   }
 	   
 		   
 	   
 	   
-	   @RequestMapping(value = {"/update" }, method = RequestMethod.POST, produces = "application/json", consumes = "application/json")
-		public void update(HttpServletResponse response, HttpServletRequest request, @RequestBody String json)
-				throws IOException {
-			AsignadorDeCharset.asignar(request, response);
-			Categoria c = (Categoria) JsonConvertidor.fromJson(json, Categoria.class);
-			categoriaDao.update(c);
-			response.getWriter().println(JsonConvertidor.toJson(c));
+	   @RequestMapping(value = {"/update/{userName}" }, method = RequestMethod.POST, produces = "application/json", consumes = "application/json")
+		public void update(HttpServletResponse response, HttpServletRequest request,
+				@RequestBody String json,  @PathVariable String userName)throws IOException {
+		   if(SesionController.verificarPermiso2(request, usuarioDao, perfilDAO, 25, sessionDao,userName)){  
+				AsignadorDeCharset.asignar(request, response);
+				Categoria c = (Categoria) JsonConvertidor.fromJson(json, Categoria.class);
+				categoriaDao.update(c);
+				response.getWriter().println(JsonConvertidor.toJson(c));
+		   }else{
+				response.sendError(403);
+			}
 		}
